+++ source/pipes/date-time-from-milliseconds.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from "@angular/core";
+import { DateTime } from "luxon";
+
+@Pipe({
+  name: "dateTimeFromMilliseconds"
+})
+export class DateTimeFromMillisecondsPipe implements PipeTransform {
+
+  transform(value: any): DateTime {
+    return DateTime.fromMillis(value);
+  }
+
+};